pyproject.toml
[project]
name = "aether-framework"
version = "0.1.0"
description = "A self-service data engineering platform."
requires-python = ">=3.9"
dependencies = [
    "pydantic",
    "pyyaml",
    "networkx",
    "typer",
    "rich",
]

[project.scripts]
aether = "aether.cli:main"

[project.optional-dependencies]
rag = [
    "sentence-transformers",
    "faiss-cpu",
    "numpy",
]
dev = [
    "pytest",
    "ruff",
    "black",
    "mypy",
    "types-PyYAML",
]
quality = [
    "pandera>=0.16.0",
    "pandas>=1.5.0",
    "great-expectations>=0.18.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C90", "N"]
ignore = ["E501"] # Ignorado porque o Black já cuida disso

[tool.ruff]
line-length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
pretty = true